+++ frontend/index.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx">
+    <script type="module" src="src/Router.tsx">
       // AWS Amplify needs this to work. See https://github.com/aws/aws-sdk-js/issues/3673
       const isBrowser = () => typeof window !== "undefined";
       const isGlobal = () => typeof global !== "undefined";
