--- frontend/src/__mocks__/styleMock.js
@@ -1,1 +1,0 @@
-module.exports = {};
