+++ frontend/src/__tests__/Button.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Button from '../components/Button/Button';
+import Button from '../views/components/Button/Button';
 
 test('renders the button with the correct label', () => {
     const handleClick = jest.fn();
