+++ frontend/src/views/components/BarChart/BarChart.css
