+++ frontend/src/views/components/BarChart/BarChart.tsx
