+++ frontend/src/views/components/Dashboard/BarChart/BarChart.tsx
