+++ frontend/src/views/components/Dashboard/BiasScore/BiasScore.tsx
