+++ frontend/src/views/components/Dashboard/GraphGrid/GraphGrid.tsx
