+++ frontend/src/views/components/Footer/Footer.tsx
