+++ frontend/src/views/components/SampleGraph/SampleGraph.css
