+++ frontend/src/views/components/SampleGraph/SampleGraph.tsx
